// ui/packages/features/workflow-builder/src/web/flow-canvas.d.ts
import "reactflow/dist/style.css";
export interface FlowCanvasProps {
    readonly nodes: readonly {
        readonly id: string;
        readonly position: {
            readonly x: number;
            readonly y: number;
        };
        readonly data: {
            readonly label: string;
        };
        readonly type: "default";
    }[];
    readonly edges: readonly {
        readonly id: string;
        readonly source: string;
        readonly target: string;
    }[];
}
export declare const FlowCanvas: import("react").NamedExoticComponent<FlowCanvasProps>;